+++ source/Public/Get-ADSyncRule.ps1
@@ -1,22 +1,56 @@
-function Get-ADSyncRule {
+function Get-ADSyncRule
+{
     [CmdletBinding(DefaultParameterSetName = 'ByName')]
-    param(
+    param (
         [Parameter(ParameterSetName = 'ByName')]
+        [Parameter(Mandatory = $true, ParameterSetName = 'ByNameAndConnector')]
         [string]
         $Name,
 
         [Parameter(ParameterSetName = 'ByIdentifier')]
         [guid]
-        $Identifier
+        $Identifier,
+
+        [Parameter(Mandatory = $true, ParameterSetName = 'ByNameAndConnector')]
+        [Parameter(Mandatory = $true, ParameterSetName = 'ByConnector')]
+        [string]
+        $ConnectorName
     )
 
-    if ($Identifier) {
+    $connectors = Get-ADSyncConnector
+
+    if ($PSCmdlet.ParameterSetName -eq 'ByIdentifier')
+    {
         ADSync\Get-ADSyncRule -Identifier $Identifier
     }
-    elseif ($Name) {
-        ADSync\Get-ADSyncRule | Where-Object Name -eq $Name
+    elseif ($PSCmdlet.ParameterSetName -eq 'ByName')
+    {
+        if ($Name)
+        {
+            ADSync\Get-ADSyncRule | Where-Object Name -EQ $Name
+        }
+        else
+        {
+            ADSync\Get-ADSyncRule
+        }
     }
-    else {
+    elseif ($PSCmdlet.ParameterSetName -eq 'ByConnector')
+    {
+        $connector = $connectors | Where-Object Name-eq $ConnectorName
+        ADSync\Get-ADSyncRule | Where-Object Connector -EQ $connector.Identifier
+    }
+    elseif ($PSCmdlet.ParameterSetName -eq 'ByNameAndConnector')
+    {
+        $connector = $connectors | Where-Object Name -EQ $ConnectorName
+        if ($null -eq $connector)
+        {
+            Write-Error "The connector '$ConnectorName' does not exist"
+            return
+        }
+        ADSync\Get-ADSyncRule | Where-Object { $_.Name -eq $Name -and $_.Connector -eq $connector.Identifier }
+    }
+    else
+    {
         ADSync\Get-ADSyncRule
     }
 }